added script scan.py

+++ workspace/tbsw.py
@@ -1,7 +1,5 @@
-#!/usr/bin/python3
-
 """
-Utility code to use TBSW in pyhton scripts 
+Some helper code to use TBSW with pyhton scripts 
 
 :author: benjamin.schwenker@phys.uni-goettinge.de  
 """
@@ -18,25 +16,26 @@ class Environment(object):
   
   :author: benjamin.schwenker@phys.uni-goettinge.de  
   """
-  def __init__(self, name, steerfiles):
+  def __init__(self, name='default', steerfiles=None):
     """
     Parameters
     ----------
     name : string 
-        name of the temporary folder holding the environment
+        name of the temporary folder holding copies of config files  
     steerfiles : string 
-        name of the folder holding steer and config files  
+        name of the folder holding config files  
     """
+    if steerfiles==None:
+      raise ValueError('Parameter steerfiles is missing')
+    
     self.name = name
-    self.runtag = ''
-    self.caltag = ''
     self.cwdir = os.getcwd() 
     self.tmpdir = os.path.join(self.cwdir+'/tmp-runs',self.name)  
     
     # create tmp-runs if not exist
     if not os.path.isdir(self.cwdir+'/tmp-runs'):
       os.mkdir(self.cwdir+'/tmp-runs')
-
+    
     # remove old tmpdir if exists 
     if os.path.isdir(self.tmpdir):
       shutil.rmtree(self.tmpdir)
@@ -46,41 +45,7 @@ class Environment(object):
       shutil.copytree(steerfiles,self.tmpdir)   
     else: 
       raise ValueError('No steerfiles can be found')
-     
-  def add_caltag(self, caltag):  
-    self.caltag = caltag
-    # check that calibration files exist
-    caldir = self.cwdir+'/cal-files/'+caltag   
-    if os.path.isdir(caldir):
-      # copy calibration files 
-      shutil.copytree(caldir,self.tmpdir+'/cal-files')  
-    else: 
-      print ('[INFO] Caltag not found') 
     
-  
-  def create_caltag(self, caltag):
-    self.caltag = caltag
-    caldir = self.cwdir+'/cal-files/'+caltag
-    
-    # create folder cal-files if not exist    
-    if not os.path.isdir(self.cwdir+'/cal-files'):
-      os.mkdir(self.cwdir+'/cal-files')
-    # overwrite caltag if exists     
-    if os.path.isdir(caldir):
-      shutil.rmtree(caldir)
-    
-    #create caltag and populate with DB files    
-    os.mkdir(caldir)		 
-    for dbfile in glob.glob(self.tmpdir + '/*DB*'): 
-      shutil.copy(dbfile, os.path.join(caldir,os.path.basename(dbfile)))  
-    
-    print ('[INFO] Created new caltag ', caltag) 
-     
-  	                           
-  def link_input(self, inputfile):
-    self.runtag = os.path.splitext(os.path.basename(inputfile))[0]
-    os.symlink( os.path.join(self.cwdir, inputfile), self.tmpdir+'/inputfilename') 
-
   def run(self,path):
     # run Marlin in tmpdir  
     os.chdir(self.tmpdir)
@@ -97,7 +62,46 @@ class Environment(object):
     
     # go back to workspace
     os.chdir(self.cwdir) 
+    
+  def add_caltag(self, caltag):  
+    # check that calibration files exist
+    caldir = self.cwdir+'/cal-files/'+caltag   
+    if os.path.isdir(caldir):
+      # copy calibration files 
+      shutil.copytree(caldir,self.tmpdir+'/cal-files')  
+    else: 
+      print ('[INFO] Caltag not found') 
+    
+  def create_caltag(self, caltag):
+    caldir = self.cwdir+'/cal-files/'+caltag
+    
+    # create folder cal-files if not exist    
+    if not os.path.isdir(self.cwdir+'/cal-files'):
+      os.mkdir(self.cwdir+'/cal-files')
+    # overwrite caltag if exists     
+    if os.path.isdir(caldir):
+      shutil.rmtree(caldir)
+    
+    #create caltag and populate with DB files    
+    os.mkdir(caldir)		 
+    for dbfile in glob.glob(self.tmpdir + '/*DB*'): 
+      shutil.copy(dbfile, os.path.join(caldir,os.path.basename(dbfile)))  
+    
+    print ('[INFO] Created new caltag ', caltag) 
+     	                           
+  def link_input(self, inputfile):
+    if os.path.isfile(inputfile):
+      os.symlink( os.path.join(self.cwdir, inputfile), self.tmpdir+'/inputfilename') 
+    else: 
+      raise ValueError('No input file found')
   
+  def get_name(self, filename):
+    localname = os.path.join(self.tmpdir,filename)
+    if os.path.isfile(localname):
+      return localname  
+    else: 
+      raise ValueError('No file found')   
+    
   def copy_rootfiles(self):
     # cd into tmpdir  
     os.chdir(self.tmpdir)
@@ -107,83 +111,98 @@ class Environment(object):
       os.mkdir(self.cwdir+'/root-files')
     
     for rootfile  in glob.glob('*.root'): 
-      name = os.path.splitext(os.path.basename(rootfile))[0]
-      shutil.move(rootfile, self.cwdir+'/root-files/'+name+'-'+self.runtag+'-'+self.caltag+'.root')  
+      basename = os.path.splitext(os.path.basename(rootfile))[0]
+      shutil.move(rootfile, self.cwdir+'/root-files/'+basename+'-'+self.name+'.root')  
     
 
-
-def simulate(steerfiles, path, caltag, ofile):
+class Simulation(Environment):
   """
-  Creates a lcio file containing N events. The events contain collections for 
-  simulated tracks measured signals. 
-  
-  :@steerfile:  name of folder containing all steering files  
-  :@path:       list containing Marlin xml files that will be executed 
-  :@caltag:     name of calibration tag (optional)
-  :@ofile:      name of output lcio file
-  
+  Class to run test beam simulations
   :author: benjamin.schwenker@phys.uni-goettinge.de  
   """
-  
-  name = os.path.splitext(os.path.basename(ofile))[0] + '-' + caltag + '-sim'
-  env = Environment(name=name, steerfiles=steerfiles)  
-  env.add_caltag(caltag)
-  env.run(path)
-  
-  src =  os.path.join(env.tmpdir, 'outputfile.slcio')	
-  dest = os.path.join(env.cwdir, ofile)	
-  shutil.move(src, dest)
-  
-  return None
-  
+  def __init__(self, steerfiles, name='sim'): 
+    Environment.__init__(self, name=name, steerfiles=steerfiles)
+    
+  def simulate(self, path=['simulation.xml'], caltag=None, ofile='mc.slcio'):
+    """
+    Creates a lcio file called ofile containing simulated events. 
+    :@path:       list containing Marlin xml files that will be executed 
+    :@caltag:     name of calibration tag (optional)
+    :@ofile:      name of output lcio file
+    
+    :author: benjamin.schwenker@phys.uni-goettinge.de  
+    """  
+    
+    print ('[INFO] Starting to simulate ' + ofile + ' ...') 
+
+    if not caltag==None:
+      self.add_caltag(caltag)
+    self.run(path)
+    
+    src =  os.path.join(self.tmpdir, 'outputfile.slcio')	
+    dest = os.path.join(self.cwdir, ofile)	
+    shutil.move(src, dest)
    
-def reconstruct(steerfiles, path, caltag, ifile):
+class Reconstruction(Environment):
   """
-  Reconstructs an input file with raw data using a caltag for calibration. 
-  
-  :@steerfile:  name of folder containing all steering files 
-  :@path:       list containing Marlin xml files that will be executed 
-  :@caltag:     name of calibration tag (optional)
-  :@ifile:      name of input file with raw data
-  
+  Class to run test beam reconstruction using calibration data
   :author: benjamin.schwenker@phys.uni-goettinge.de  
-  """   
-  
-  print ('[INFO] Starting to process file ' + ifile + ' ...')  
-  
-  name = os.path.splitext(os.path.basename(ifile))[0] + '-' + caltag + '-reco'
-  env = Environment(name=name, steerfiles=steerfiles)  
-  env.add_caltag(caltag)
-  env.link_input(ifile)
-  env.run(path)
-  env.copy_rootfiles()
-
-  print ('[INFO] Done processing file ' + ifile)  
-
-  return None
-
-
-def calibrate(steerfiles, path, caltag, ifile):
   """
-  Calibrate raw data from a test beam experiment. 
+  def __init__(self, steerfiles, name='reco'): 
+    Environment.__init__(self, name=name, steerfiles=steerfiles)
   
-  :@steerfile:  name of folder containing all steering files 
-  :@marlinpath: list containing Marlin xml files that will be executed 
-  :@caltag:     name of calibration tag (optional)
-  :@ifile:      name of input file with raw data
-  
-  :author: benjamin.schwenker@phys.uni-goettinge.de  
-  """   
-  
-  print ('[INFO] Starting to process file ' + ifile + ' ...')    
-  
-  name = os.path.splitext(os.path.basename(ifile))[0] + '-' + caltag + '-cal'
-  env = Environment(name=name, steerfiles=steerfiles)  
-  env.link_input(ifile)
-  env.run(path)
-  env.create_caltag(caltag) 
-  
-  print ('[INFO] Done processing file ' + ifile)  
+  def reconstruct(self, path=['reco.xml'], caltag=None, ifile=None):
+    """
+    Reconstructs an input file with raw data using a caltag for calibration. 
+    :@path:       list containing Marlin xml files that will be executed 
+    :@caltag:     name of calibration tag (optional)
+    :@ifile:      name of input file with raw data
+    
+    :author: benjamin.schwenker@phys.uni-goettinge.de  
+    """  
+    
+    if ifile==None:
+      raise ValueError('Parameter ifile is missing') 
+
+    if caltag==None:
+      raise ValueError('Parameter caltag is missing') 
+    
+    print ('[INFO] Starting to reconstruct file ' + ifile + ' ...')  
      
-  return None
+    self.add_caltag(caltag)
+    self.link_input(ifile)
+    self.run(path)
+    self.copy_rootfiles()
+    
+    print ('[INFO] Done processing file ' + ifile)  
 
+
+class Calibration(Environment):
+  """
+  Class to run test beam calibration using calibration run
+  :author: benjamin.schwenker@phys.uni-goettinge.de  
+  """
+  def __init__(self, steerfiles, name='cal'): 
+    Environment.__init__(self, name=name, steerfiles=steerfiles)
+  
+  def calibrate(self, path=[], caltag='default', ifile=None):
+    """
+    Calibrate beam telescope using a calibration run  
+    :@path: list containing Marlin xml files that will be executed 
+    :@caltag:     name of calibration tag (optional)
+    :@ifile:      name of input file with raw data
+    
+    :author: benjamin.schwenker@phys.uni-goettinge.de  
+    """   
+
+    if ifile==None:
+      raise ValueError('Parameter ifile is missing') 
+    
+    print ('[INFO] Starting to calibrate file ' + ifile + ' ...')    
+    
+    self.link_input(ifile)
+    self.run(path)
+    self.create_caltag(caltag) 
+    
+    print ('[INFO] Done processing file ' + ifile)  
+
